+++ src/realm/sync/client.hpp
@@ -515,11 +515,6 @@ public:
         /// application can listen to change notifications from the client
         /// reset exactly as in a DOWNLOAD message.
         ///
-        /// The client reset will recover non-uploaded changes in the local
-        /// Realm if and only if 'recover_local_changes' is true. In case,
-        /// 'recover_local_changes' is false, the local Realm state will hence
-        /// be set to the server's state (server wins).
-        ///
         /// Async open and client reset require a private directory for
         /// metadata. This directory must be specified in the option
         /// 'metadata_dir'. The metadata_dir must not be touched during async
@@ -568,22 +563,8 @@ public:
         /// void async_wait_for_download_completion(WaitOperCompletionHandler)
         /// or
         /// bool wait_for_download_complete_or_client_stopped().
-        ///
-        /// The option 'require_recent_state_realm' is used for async open to
-        /// request a recent state Realm. A recent state Realm is never empty
-        /// (unless there is no data), and is recent in the sense that it was
-        /// produced by the current incarnation of the server. Recent does not
-        /// mean the absolutely newest possible state Realm, since that might
-        /// lead to too excessive work on the server. Setting
-        /// 'require_recent_state_realm' to true might lead to more work
-        /// performed by the server but it ensures that more data is downloaded
-        /// using async open instead of ordinary synchronization. It is
-        /// recommended to set 'require_recent_state_realm' to true. Client
-        /// reset always downloads a recent state Realm.
         struct ClientReset {
             std::string metadata_dir;
-            bool recover_local_changes = true;
-            bool require_recent_state_realm = true;
         };
         util::Optional<ClientReset> client_reset_config;