+++ CHANGELOG.md
@@ -22,6 +22,11 @@
 -----------
 
 ### Internals
+* Remove ClientResyncMode::Recover option. This doesn't work and shouldn't be used by anyone currently.
+* Remove sync support for CLIENT_VERSION request; it isn't used.
+* Remove the STATE message request and handling from the client and test sever. The current server was always responding with empty state anyhow. Clients will now consider receiving this message as an error because they never request it.
+* Handling for async open, involving state files is mostly removed.
+* When a client reset occured, we would create a metadata Realm to handle state download but this isn't needed anymore so it is removed.
 * Realm::write_copy_without_client_file_id and Realm::write_copy are merged. For synchronized realms, the file written will have the client file ident removed. Furthermore it is required that all local changes are synchronized with the server before the copy can. The function will throw if there are pending uploads.
 
 ----------------------------------------------